avoid mangling/including tmp.XXXXXXXX in tar base path
As it was, the base path for extracted files was ending up tmtmp.XXXXXXXX because of the xform param passed to tar command.

Complements the previous change which uses unique temp dir for each run.  This is especially important since this version of script is being used for OSD tiered access where the "current" dir would naturally be /root.